+++ packages/cli/src/Site/SiteArtifactRendererFactory.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Waaseyaa\CLI\Site;
+
+use Waaseyaa\CLI\Site\Recipe\PublishedContentRecipe;
+use Waaseyaa\SiteContract\Generation\SiteArtifactRenderer;
+
+final class SiteArtifactRendererFactory
+{
+    public static function create(): SiteArtifactRenderer
+    {
+        return new SiteArtifactRenderer([
+            new PublishedContentRecipe(),
+        ]);
+    }
+}